refactor: blog wrapper 컴포넌트 생성 및 prefetch true 설정
