+++ features/blog/index.ts
@@ -1,3 +1,4 @@
 export { BlogIntroduce } from './BlogIntroduce';
 export { BlogTags } from './BlogTags';
 export { BlogPosts } from './BlogPosts';
+export { BlogPostWrapper } from './BlogPostWrapper';
